+++ tools/shared.py
@@ -992,9 +992,6 @@ def verify_settings():
     if Settings.CYBERDWARF:
       exit_with_error('emcc: CYBERDWARF is not supported by the LLVM wasm backend')
 
-    if Settings.EMTERPRETIFY:
-      exit_with_error('emcc: EMTERPRETIFY is not supported by the LLVM wasm backend')
-
     if Settings.EMULATED_FUNCTION_POINTERS:
       exit_with_error('emcc: EMULATED_FUNCTION_POINTERS is not meaningful with the wasm backend.')