+++ ChangeLog.md
@@ -39,6 +39,9 @@ Current Trunk
 - Remove `BINARYEN_PASSES` settings.  We still have `BINARYEN_EXTRA_PASSES`, but
   completely overriding the set of passes from the command line didn't make much
   sense.
+- Remove emterpreter and ``EMTERPRETIFY`` settings.  Emterpreter has largerly
+  been replaces by asyncify and is fastcomp only so due for removing in
+  the near future anyway.
 
 v1.39.13: 04/17/2020
 --------------------